fix: address dense union review feedback
