+++ encodings/dense-union/Cargo.toml
@@ -16,14 +16,15 @@ version = { workspace = true }
 [dependencies]
 prost = { workspace = true }
 vortex-array = { workspace = true }
+vortex-buffer = { workspace = true }
 vortex-error = { workspace = true }
 vortex-mask = { workspace = true }
 vortex-session = { workspace = true }
 
 [dev-dependencies]
 divan = { workspace = true }
+rstest = { workspace = true }
 vortex-array = { workspace = true, features = ["_test-harness"] }
-vortex-buffer = { workspace = true }
 
 [lints]
 workspace = true
