+++ encodings/dense-union/src/canonical.rs
@@ -11,6 +11,8 @@ use vortex_array::arrays::Primitive;
 use vortex_array::arrays::PrimitiveArray;
 use vortex_array::arrays::UnionArray;
 use vortex_array::scalar::Scalar;
+use vortex_array::validity::Validity;
+use vortex_buffer::BufferMut;
 use vortex_error::VortexResult;
 use vortex_error::vortex_ensure;
 use vortex_error::vortex_err;
@@ -20,6 +22,16 @@ use crate::array::DenseUnion;
 use crate::array::DenseUnionArrayExt;
 use crate::array::DenseUnionArraySlotsExt;
 
+/// Converts a dense union to its canonical sparse representation.
+///
+/// Each child is a dictionary over the original compact child, so values are not copied. Codes for
+/// other variants stay zero because their type IDs make them unreachable. Unused variants use a
+/// constant zero code array, and empty children use a one-value constant because dictionaries
+/// require non-empty values.
+///
+/// # Errors
+///
+/// Returns an error for unknown type IDs, invalid offsets, or failed array construction.
 pub(crate) fn canonicalize(
     array: Array<DenseUnion>,
     ctx: &mut ExecutionCtx,
@@ -31,29 +43,31 @@ pub(crate) fn canonicalize(
     let type_id_values = type_ids.as_slice::<u8>();
     let offset_values = offsets.as_slice::<i32>();
     let valid_rows = type_ids.validity()?.execute_mask(len, ctx)?;
-    let mut codes_by_child = vec![vec![0u32; len]; variants.len()];
+    let child_lengths = array.iter_children().map(ArrayRef::len).collect::<Vec<_>>();
+
+    let mut child_indices = [None; 256];
+    for (child_index, type_id) in variants.type_ids().iter().copied().enumerate() {
+        child_indices[usize::from(type_id)] = Some(child_index);
+    }
+    let mut codes_by_child: Vec<Option<BufferMut<u32>>> = vec![None; variants.len()];
 
     let mut assign_row = |row: usize| -> VortexResult<()> {
         let type_id = type_id_values[row];
-        let child_index = variants
-            .tag_to_child_index(type_id)
+        let child_index = child_indices[usize::from(type_id)]
             .ok_or_else(|| vortex_err!("DenseUnion contains unknown type ID {type_id}"))?;
-        let offset = usize::try_from(offset_values[row]).map_err(|_| {
+        let offset = u32::try_from(offset_values[row]).map_err(|_| {
             vortex_err!(
                 "DenseUnion contains negative offset {} at row {row}",
                 offset_values[row]
             )
         })?;
-        let child_len = array
-            .child(child_index)
-            .ok_or_else(|| vortex_err!("DenseUnion is missing compact child {child_index}"))?
-            .len();
+        let child_len = child_lengths[child_index];
         vortex_ensure!(
-            offset < child_len,
+            (offset as usize) < child_len,
             "DenseUnion offset {offset} is out of bounds for child {child_index} of length {child_len}"
         );
-        codes_by_child[child_index][row] = u32::try_from(offset)
-            .map_err(|_| vortex_err!("DenseUnion offset {offset} does not fit in u32"))?;
+        let codes = codes_by_child[child_index].get_or_insert_with(|| BufferMut::zeroed(len));
+        codes[row] = offset;
         Ok(())
     };
 
@@ -75,13 +89,18 @@ pub(crate) fn canonicalize(
         .iter_children()
         .zip(codes_by_child)
         .map(|(child, codes)| {
+            let codes = match codes {
+                Some(codes) => {
+                    PrimitiveArray::new(codes.freeze(), Validity::NonNullable).into_array()
+                }
+                None => ConstantArray::new(0u32, len).into_array(),
+            };
             let values = if child.is_empty() {
                 ConstantArray::new(Scalar::default_value(child.dtype()), 1).into_array()
             } else {
                 child.clone()
             };
-            DictArray::try_new(PrimitiveArray::from_iter(codes).into_array(), values)
-                .map(IntoArray::into_array)
+            DictArray::try_new(codes, values).map(IntoArray::into_array)
         })
         .collect::<VortexResult<Vec<_>>>()?;
